+++ HISTORY.md
@@ -2,14 +2,28 @@
 
 ---
 
+### 2.0.0
+
+`tag:new` 重量级功能，增加对多级下拉的支持。
+
+`tag:new` 对 model 增加 disabled 字段，支持使某个选项设置为不可选。
+
+`tag:fixed` DOM 中 `data-defaultSelected` 属性纠正为 `data-default-selected`，用户传入的参数依然是 `defaultSelected`。
+
+`tag:changed` selected 和 hover 状态的 className 由 `{{classPrefix}}-selected` 和 `{{classPrefix}}-hover` 替换为更加统一的 `{{classPrefix}}-item-selected` 和 `{{classPrefix}}-item-hover`
+
 ### 0.9.0 [milestone](https://github.com/aralejs/select/issues?milestone=1&page=1&state=closed)
 
-- `[feature]` [#4](https://github.com/aralejs/select/issues/4) 支持操作 option
-- `[feature]` [#8](https://github.com/aralejs/select/issues/8) 支持表单项
-- `[feature]` [#9](https://github.com/aralejs/select/issues/9) trigger 和浮层应该等宽
-- `[feature]` [#11](https://github.com/aralejs/select/issues/11) firefox css('border') 失效
-- `[feature]` [#15](https://github.com/aralejs/select/issues/15) destroy的bug
+`tag:new` [#4](https://github.com/aralejs/select/issues/4) 支持操作 option
+
+`tag:new` [#8](https://github.com/aralejs/select/issues/8) 支持表单项
+
+`tag:improved` [#9](https://github.com/aralejs/select/issues/9) trigger 和浮层应该等宽
+
+`tag:fixed` [#11](https://github.com/aralejs/select/issues/11) firefox css('border') 失效
+
+`tag:fixed` [#15](https://github.com/aralejs/select/issues/15) destroy的bug
 
 ### 0.8.0
 
-- `[feature]` 初始化
+`tag:new` 初始化